api-gateway/requirements.txt
flask
python-dotenv
pyjwt
requests